create moule  to attach product image directly from the internet
This Cybrosys's module allows you to easily attach product image directly from the internet.
You don't need to bother the product for an image.
Just Google the product's name and identify & select it's photo. Copy the image URL and paste in image URL field in Odoo.
This module will attach the corresponding photo in product form.